v1.17.4 * fixed ratings to use ints

+++ app/src/main/java/com/insertcoolnamehere/showandsell/GroupDetailActivity.java
@@ -170,7 +170,7 @@ public class GroupDetailActivity extends AppCompatActivity implements BrowseFrag
                             .appendPath("groups")
                             .appendPath("rategroup")
                             .appendQueryParameter("id", mGroupId)
-                            .appendQueryParameter("rating", rating+"")
+                            .appendQueryParameter("rating", ((int)rating)+"")
                             .appendQueryParameter("userId", userId)
                             .appendQueryParameter("password", password)
                             .build().toString();
